+++ src/main/java/nextstep/favorite/application/FavoriteService.java
@@ -5,7 +5,6 @@ import nextstep.common.exception.*;
 import nextstep.favorite.domain.Favorite;
 import nextstep.favorite.infrastructure.FavoriteRepository;
 import nextstep.favorite.presentation.FavoriteRequest;
-import nextstep.favorite.presentation.FavoriteResponse;
 import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.infrastructure.MemberRepository;
