+++ src/main/java/nextstep/favorite/presentation/FavoriteController.java
@@ -1,6 +1,7 @@
 package nextstep.favorite.presentation;
 
 import lombok.RequiredArgsConstructor;
+import nextstep.favorite.application.FavoriteResponse;
 import nextstep.favorite.application.FavoriteService;
 import nextstep.member.domain.LoginMember;
 import nextstep.member.presentation.AuthenticationPrincipal;
