+++ src/main/java/nextstep/member/application/GithubClient.java
@@ -1,7 +1,6 @@
 package nextstep.member.application;
 
 import nextstep.member.infrastructure.GithubProfileResponse;
-import nextstep.member.presentation.TokenResponse;
 
 public interface GithubClient {
     TokenResponse getAccessTokenFromGithub(String code);
