+++ src/main/java/nextstep/member/application/TokenService.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 import nextstep.member.AuthenticationException;
 import nextstep.member.domain.Member;
 import nextstep.member.infrastructure.MemberRepository;
-import nextstep.member.presentation.TokenResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
