+++ src/main/java/nextstep/subway/application/StationResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.presentation;
+package nextstep.subway.application;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
