+++ src/test/java/nextstep/subway/acceptance/LineAcceptanceTest.java
@@ -3,7 +3,7 @@ package nextstep.subway.acceptance;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.presentation.LineRequest;
-import nextstep.subway.presentation.LineResponse;
+import nextstep.subway.application.LineResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
